% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_regarima.R
\name{regarimaoutliers}
\alias{regarimaoutliers}
\title{Title}
\usage{
regarimaoutliers(
  y,
  order = c(0L, 1L, 1L),
  seasonal = c(0L, 1L, 1L),
  mean = F,
  X = NULL,
  X.td = NULL,
  ao = T,
  ls = T,
  tc = F,
  so = F,
  cv = 0
)
}
\arguments{
\item{y}{Series}

\item{order}{Regular orders (p,d,q)}

\item{seasonal}{Seasonal orders (bp, bd, bq)}

\item{mean}{Mean correction}

\item{X}{Regression variables}

\item{X.td}{Trading days ( groups of days: for instance X.td=c(1,1,1,1,1,2,0))}

\item{ao}{Detection of additive outliers}

\item{ls}{Detection of level shifts}

\item{tc}{Detection of transitory changes}

\item{so}{Detection of seasonal outliers}

\item{cv}{Critical value}
}
\value{

}
\description{
Title
}
