% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stsoutliers.R
\name{stsoutliers}
\alias{stsoutliers}
\title{Title}
\usage{
stsoutliers(
  y,
  level = 1,
  slope = 1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  X = NULL,
  X.td = NULL,
  ao = TRUE,
  ls = TRUE,
  so = FALSE,
  cv = 0,
  tcv = 0,
  estimation.forward = c("Score", "Point", "Full"),
  estimation.backward = c("Point", "Score", "Full")
)
}
\arguments{
\item{estimation.backward}{}
}
\description{
Title
}
